requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
minio==7.1.0
pydantic==1.9.0
jupyter==1.0.0
requests==2.27.1
mysql-connector-python==8.0.29
supabase-py==0.2.8
python-dotenv==0.20.0